SyncConfigs: include reading local configs functionality
Moved this responsibility from JoinData
