+++ Source/Client/Util/SyncConfigs.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using System.Diagnostics.CodeAnalysis;
 using System.IO;
 using System.Linq;
 using System.Reflection;
@@ -12,7 +13,7 @@ namespace Multiplayer.Client.Util;
 /// Responsible for saving a server's config files and retrieving them later.
 public static class SyncConfigs
 {
-    private static readonly string TempConfigsPath = GenFilePaths.FolderUnderSaveData(JoinData.TempConfigsDir);
+    private static readonly string TempConfigsPath = GenFilePaths.FolderUnderSaveData("MultiplayerTempConfigs");
     private const string RestartConfigsVariable = "MultiplayerRestartConfigs";
 
     public static bool Applicable { private set; get; }
@@ -34,6 +35,85 @@ public static class SyncConfigs
 
         foreach (var config in configs)
             File.WriteAllText(Path.Combine(TempConfigsPath, $"Mod_{config.ModId}_{config.FileName}.xml"), config.Contents);
+    }
+
+    [SuppressMessage("ReSharper", "StringLiteralTypo")]
+    public static string[] ignoredConfigsModIds =
+    [
+        // todo unhardcode it
+        "rwmt.multiplayer",
+        "hodlhodl.twitchtoolkit", // contains username
+        "dubwise.dubsmintmenus",
+        "dubwise.dubsmintminimap",
+        "arandomkiwi.rimthemes",
+        "brrainz.cameraplus",
+        "giantspacehamster.moody",
+        "fluffy.modmanager",
+        "jelly.modswitch",
+        "betterscenes.rimconnect", // contains secret key for streamer
+        "jaxe.rimhud",
+        "telefonmast.graphicssettings",
+        "derekbickley.ltocolonygroupsfinal",
+        "dra.multiplayercustomtickrates", // syncs its own settings
+        "merthsoft.designatorshapes" // settings for UI and stuff meaningless for MP
+        //"zetrith.prepatcher",
+    ];
+
+    public const string HugsLibId = "unlimitedhugs.hugslib";
+    public const string HugsLibSettingsFile = "ModSettings";
+
+    public static List<ModConfig> GetSyncableConfigContents(List<string> modIds)
+    {
+        var list = new List<ModConfig>();
+
+        foreach (var modId in modIds)
+        {
+            if (ignoredConfigsModIds.Contains(modId)) continue;
+
+            var mod = LoadedModManager.RunningMods.FirstOrDefault(m =>
+                m.PackageIdPlayerFacing.ToLowerInvariant() == modId);
+            if (mod == null) continue;
+
+            foreach (var modInstance in LoadedModManager.runningModClasses.Values)
+            {
+                if (modInstance.modSettings == null) continue;
+                if (!mod.assemblies.loadedAssemblies.Contains(modInstance.GetType().Assembly)) continue;
+
+                var instanceName = modInstance.GetType().Name;
+
+                // This path may point to configs downloaded from the server
+                var file = LoadedModManager.GetSettingsFilename(mod.FolderName, instanceName);
+
+                if (File.Exists(file))
+                    list.Add(GetConfigCatchError(file, modId, instanceName));
+            }
+        }
+
+        // Special case for HugsLib
+        if (modIds.Contains(HugsLibId) && JoinData.GetInstalledMod(HugsLibId) is { Active: true })
+        {
+            var hugsConfig = HugsLib_OverrideConfigsPatch.HugsLibConfigIsOverriden
+                ? HugsLib_OverrideConfigsPatch.HugsLibConfigOverridePath
+                : Path.Combine(GenFilePaths.SaveDataFolderPath, "HugsLib", "ModSettings.xml");
+
+            if (File.Exists(hugsConfig))
+                list.Add(GetConfigCatchError(hugsConfig, HugsLibId, HugsLibSettingsFile));
+        }
+
+        return list;
+
+        ModConfig GetConfigCatchError(string path, string id, string file)
+        {
+            try
+            {
+                return new ModConfig(id, file, Contents: File.ReadAllText(path));
+            }
+            catch (Exception e)
+            {
+                Log.Error($"Exception getting config contents {file}: {e}");
+                return new ModConfig(id, "ERROR", "");
+            }
+        }
     }
 
     public static string GetConfigPath(string modId, string handleName) =>
@@ -64,7 +144,7 @@ static class OverrideConfigsPatch
         if (mod == null)
             return;
 
-        if (JoinData.ignoredConfigsModIds.Contains(mod.ModMetaData.PackageIdNonUnique))
+        if (SyncConfigs.ignoredConfigsModIds.Contains(mod.ModMetaData.PackageIdNonUnique))
             return;
 
         __result = SyncConfigs.GetConfigPath(mod.PackageIdPlayerFacing.ToLowerInvariant(), modHandleName);
@@ -75,8 +155,9 @@ static class OverrideConfigsPatch
 [HarmonyPatch]
 static class HugsLib_OverrideConfigsPatch
 {
-    public static string HugsLibConfigOverridenPath =
-        SyncConfigs.GetConfigPath(JoinData.HugsLibId, JoinData.HugsLibSettingsFile);
+    public static string HugsLibConfigOverridePath =
+        SyncConfigs.GetConfigPath(SyncConfigs.HugsLibId, SyncConfigs.HugsLibSettingsFile);
+    public static bool HugsLibConfigIsOverriden => File.Exists(HugsLibConfigOverridePath);
 
     private static readonly MethodInfo MethodToPatch =
         AccessTools.Method("HugsLib.Core.PersistentDataManager:GetSettingsFilePath");
@@ -89,7 +170,7 @@ static class HugsLib_OverrideConfigsPatch
     {
         if (!SyncConfigs.Applicable) return;
         if (__instance.GetType().Name != "ModSettingsManager") return;
-        if (!File.Exists(HugsLibConfigOverridenPath)) return;
-        __instance.SetPropertyOrField("OverrideFilePath", HugsLibConfigOverridenPath);
+        if (!HugsLibConfigIsOverriden) return;
+        __instance.SetPropertyOrField("OverrideFilePath", HugsLibConfigOverridePath);
     }
 }
